requirements.txt
langchain
langchain_community
PyPDF2==3.0.1
qdrant-client==1.11.3
requests==2.31.0
tqdm==4.66.1
google-generativeai>=0.8.0
sentence-transformers>=3.0.0